refactor(api): move GET /api/users/me to deprecated context
