+++ api/src/deprecated/application/routes.js
@@ -1,9 +1,10 @@
 import { prescriberInformationsRoute } from './prescriber-informations.route.js';
+import { usersMeRoute } from './users-me.route.js';
 
 const register = async function (server) {
-  server.route([...prescriberInformationsRoute]);
+  server.route([...prescriberInformationsRoute, ...usersMeRoute]);
 };
 
-const name = 'deprecated/prescriber-api';
+const name = 'deprecated/deprecated-api';
 
 export const deprecatedRoutes = [{ register, name }];
