+++ api/src/deprecated/domain/models/UserWithActivity.js
