+++ api/src/deprecated/infrastructure/serializers/jsonapi/user-with-activity.serializer.js
