+++ api/tests/deprecated/integration/domain/usecases/get-prescriber.test.js
