+++ api/tests/deprecated/unit/domain/models/UserWithActivity_test.js
@@ -1,13 +1,13 @@
 import sinon from 'sinon';
 
+import { UserWithActivity } from '../../../../../src/deprecated/domain/models/UserWithActivity.js';
 import { User } from '../../../../../src/identity-access-management/domain/models/User.js';
-import { UserWithActivity } from '../../../../../src/identity-access-management/domain/read-models/UserWithActivity.js';
 import { STATUS } from '../../../../../src/legal-documents/domain/models/LegalDocumentStatus.js';
 import { config } from '../../../../../src/shared/config.js';
 import { expect } from '../../../../test-helper.js';
 import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
 
-describe('Unit | Domain | Read-Models | UserWithActivity', function () {
+describe('Deprecated | Unit | Domain | Read-Models | UserWithActivity', function () {
   describe('constructor', function () {
     context('when the user has accepted the CGU', function () {
       it('returns the user with accepted TOS status', async function () {
